third_party/WebKit/LayoutTests/imported/wpt/uievents/throwing-in-listener-and-window-error-event-expected.txt

CONSOLE ERROR: line 22: Uncaught Error: Error from listener
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: Error from listener
Harness: the test ran to completion.
